Extract HTTP responses
